tabs_generator/pyproject.toml
[project]
name = "guitar-player-tabs-generator"
version = "0.1.0"
description = "Guitar tab transcription service using basic-pitch"
requires-python = ">=3.10"
dependencies = [
    "setuptools<81",
    "basic-pitch[onnx]",
    "numpy>=1.24.0,<2",
    "librosa>=0.10.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.35.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "mangum>=0.19.0",
    "python-json-logger>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tabs_generator"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
extraPaths = ["src"]
